creacion de postSlide
